Refactored classifying function as parameter.
Classifying Function - The function responsible
for sampling the search space, returns true/false
for the point sampled was of target performance.

Instead of a struct field, it is now a paramter
into the step() and sample_next() methods. This
simplifies implementation a bit, and allows for
closures to be used.

The Classifier type is defined as
    Box<dyn Fn(SVector<f64, N>)
        -> Result<bool, SamplingError>