Refactored struct PointNode as enum Sample
This was to better support type system. Allows for
compile-time errors for when a function depends on
a target vs non-target sample.

Also, highly improved MeshExplorer step method.
Less code, simpler, doesn't involve unnecessary
adherer.take() and reduced redundancy.

Not a fan of const_adherer at line 101. Future
changes to this shouldn't effect anything so it
can be delayed.