First attempt at comp-time sample perf checking.
This is not as clean as I would like it, leading
to the branch. Samples are not being treated as if
they were vectors, and instead must be explicitly
dereferenced (e.g. *my_sample). Looks awful,
doesn't read well, and has a lot of additional
struct and impl definitions.

Overall, may revert.